pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maxllm"
version = "0.1.0"
description = "A unified OpenAI API client with rate limiting, caching, and batch processing"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "tqdm>=4.65.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "aiolimiter>=1.1.0",
    "diskcache>=5.6.0",
    "json5>=0.9.0",
    "json-repair>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
transformers = [
    "transformers>=4.30.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/maxllm"
Repository = "https://github.com/yourusername/maxllm"
Issues = "https://github.com/yourusername/maxllm/issues"

[tool.setuptools]
packages = ["maxllm"]

[tool.setuptools.package-data]
# Explicitly exclude litellm.yaml from the package
"*" = []
